+++ src/Entity/Preorder.php
@@ -27,7 +27,7 @@ class Preorder
     private ?\DateTimeInterface $ordered_at = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $price = null;
+    private ?float $price = null;
 
     #[ORM\Column(type: Types::DATE_MUTABLE, nullable: true)]
     private ?\DateTimeInterface $sent_at = null;
@@ -85,12 +85,12 @@ class Preorder
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(?int $price): self
+    public function setPrice(?float $price): self
     {
         $this->price = $price;
